services/workers/requirements.txt
celery==5.3.6
redis==5.0.1
kafka-python==2.0.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
opensearch-py==2.4.2
neo4j==5.16.0
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20
httpx==0.26.0
beautifulsoup4==4.12.3
structlog==24.1.0
python-dateutil==2.8.2
tenacity==8.2.3
